+++ modules/pcf-exchange/src/main/java/org/eclipse/tractusx/sde/pcfexchange/service/impl/AsyncPushPCFDataForApproveRequest.java
@@ -96,7 +96,7 @@ public class AsyncPushPCFDataForApproveRequest {
 						// push api call
 						Runnable runnable = () -> proxyRequestInterface.sendNotificationToConsumer(status,
 								calculatedPCFValue, request.getProductId(), request.getBpnNumber(),
-								request.getRequestId(), request.getMessage());
+								request.getRequestId(), request.getMessage(), false);
 
 						new Thread(runnable).start();