+++ modules/sde-external-services/digital-twins/src/main/java/org/eclipse/tractusx/sde/digitaltwins/entities/common/ProtocolInformation.java
@@ -1,7 +1,7 @@
 /********************************************************************************
  * Copyright (c) 2022 BMW GmbH
- * Copyright (c) 2022, 2023 T-Systems International GmbH
- * Copyright (c) 2022, 2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2022,2024 T-Systems International GmbH
+ * Copyright (c) 2022,2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -42,7 +42,7 @@ public class ProtocolInformation {
     private String endpointAddress;
     private String endpointProtocol;
     private List<String> endpointProtocolVersion;
-    private String subProtocol;
+    private String subprotocol;
     private String subprotocolBody;
     private String subprotocolBodyEncoding;
     private List<SecurityAttributes> securityAttributes;